Highlight markers without blue lines
Adds an option to use pink markers when a location doesn't have a blue
line yet, useful for finding new coverage as it releases. Pink because
it's also what sv-map uses for new lines :)

It works by loading tiles from the blue lines layer and checking the
pixel at the location where the marker is. If it's not fully
transparent, there's a line there.

This slows down generation a bit because there's an extra network
request involved so it's off by default.